daemon: Move WebAuthnError out of common
