+++ credentialsd/src/gateway/util.rs
@@ -2,7 +2,6 @@
 //!
 //! Types shared between components within this service belong in credentialsd_common::model.
 
-use credentialsd_common::model::WebAuthnError;
 use libwebauthn::ops::webauthn::idl::origin::{
     Origin as LibwebauthnOrigin, RequestOrigin as LibwebauthnRequestOrigin,
 };
@@ -12,13 +11,15 @@ use libwebauthn::ops::webauthn::{
     ReqwestRelatedOriginsSource,
 };
 
-use crate::gateway::{GetCredentialRequest, GetPublicKeyCredentialResponse};
-use crate::webauthn::{
-    GetAssertionRequest, MakeCredentialRequest, NavigationContext, Origin, WebAuthnIDLResponse,
-};
 use crate::{
-    gateway::{CreateCredentialRequest, CreatePublicKeyCredentialResponse},
+    gateway::{
+        CreateCredentialRequest, CreatePublicKeyCredentialResponse, GetCredentialRequest,
+        GetPublicKeyCredentialResponse, WebAuthnError,
+    },
     model::{GetAssertionResponseInternal, MakeCredentialResponseInternal},
+    webauthn::{
+        GetAssertionRequest, MakeCredentialRequest, NavigationContext, Origin, WebAuthnIDLResponse,
+    },
 };
 
 impl TryFrom<&Origin> for LibwebauthnOrigin {
